Dashboard menu - code update
Moved DB code to dashboard_model
